fixup rebase problems

+++ lib/device/NDevice/NDevice.cpp
@@ -92,7 +92,7 @@ NDevice::~NDevice()
 bool NDevice::processCommand(const FUJI_COMMAND_PACKET &packet)
 {
     {
-        uint8_t cmd = packet.command();
+        uint8_t cmd = (uint8_t) packet.command();
         Debug_printf("NDevice=0x%02x processCommand: '%c' 0x%02x\n",
                      packet.device(), isprint(cmd) ? cmd : '.', cmd);
     }
